Fix DuplicateTestNameInterceptor
